Deal properly with provided factor levels intended to be mapped to NA.
This exploits the feature of readr::fct_recode that removes levels named NULL.

My understanding is that it is not possible to have elements of a vector equal to NULL, so I could not handle this in the same way as the other arguments to fct_recode, i.e. by !!!ing recode_vector. So instead have modified the fct_recode call after it is created, by changing the arg names from "" to "NULL".