added a new feature to test_make_spell. this is  admission_method_type which is added since the new functions that will develop overtime requires more columns addited to the make_spell_table.
For example, the admission_method_type was not part of the spell_table until I developed the new function plot - los_distribution_method_admission. Yet, it required me to go back to the spell_linkage function and add a new column - that of admission method type. With the test developed for los_distrib_method_admission the spell_linkage test failed. Thus needed to add this new feature to the test so that it does not break again. It is a natural behavior with the additions of new columns to the spell table.